[Webkit] Add emacs-xwidget-internal variable, default nil
The new versions of libwebkit2gtk is causing emacs crash when calling any webkit
functions. There are many threads in emacs mailing lists reporting this bug and
a conclusive solution still in progress.

For the facts mentioned above, I created a new variable that toggles the using
of xwidget-webkit(even if its available) since in some environments still works.
With this, the troubleshooting is less painful and if some issue appears just
disable it and the browser is used.

I hope that emacs devs find a solution for this, webkit is nice feature for the
editor :)