[Corfu] Introduce corfu package as completion framework option
Add corfu package configuration as another completion framework option, compared
with company(the default framework choice), its more lightweight and fast. Also
requires much less configuration.

The variable 'emacs-completion-framework' controls which package must be used
and the default value is company and if others packages that integrate with
company must be installed or not.

Also fix some warnings in interface.