[Tree-sitter] Choose between build-in and third-party package
Since emacs 29.1, tree-sitter was merged and can be enabled with the flag
"--with-tree-sitter", the lib is called "treesit" and does a great job. But, the
third-party package "tree-sitter" does a better job highlighting. It's possible
to do the same with treesit, but needs a config for every language, which
requires much more work.

For the reasons above I created a new option, "emacs-parsing-system", which
controls which parsing system has to be used by emacs, the default is
"tree-sitter".

All the packages with "ts-mode" now are only installed when the built-in option
is used.

For tree-sitter config the function "silence-function-messages" was added to
prevent annoying messages incoming from tree-sitter-lags, which in future will
support the built-in "ts-mode".